requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
distro==1.8.0
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==4.2.7
django-cache-url==3.4.4
django-environ==0.11.2
environs==9.5.0
exceptiongroup==1.1.3
frozenlist==1.4.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
lxml==4.9.3
marshmallow==3.20.1
multidict==6.0.4
openai==0.28.1
packaging==23.2
Pillow==10.1.0
psycopg==3.1.8
psycopg-binary==3.1.8
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-pptx==0.6.23
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
whitenoise==6.4.0
XlsxWriter==3.1.9
yarl==1.9.2
